Alterações no tipo de dado da PK, implementação do BairroRepository

+++ src/main/java/com/controle_de_gastos/notas_api/Repository/AtributoRepository.java
@@ -3,5 +3,5 @@ package com.controle_de_gastos.notas_api.Repository;
 import com.controle_de_gastos.notas_api.model.Atributo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AtributoRepository extends JpaRepository<Atributo, Long> {
+public interface AtributoRepository extends JpaRepository<Atributo, Integer> {
 }
